+++ build.ps1
@@ -1,5 +1,5 @@
 # Build script to create Zotero plugin XPI file
-# Publication Rankings Plugin for Zotero 7
+# Publication Rankings Plugin for Zotero 7+
 #
 # Copyright (C) 2025 Ben Stephens
 # Licensed under GNU General Public License v3.0 (GPLv3)
@@ -7,7 +7,7 @@
 # XPI files are just ZIP files with a different extension
 
 $pluginName = "publication-rankings"
-$version = "0.3.0"
+$version = "0.3.1"
 $outputFile = "$pluginName-$version.xpi"
 
 # Remove old XPI if it exists
@@ -105,7 +105,7 @@ Remove-Item $tempDir -Recurse -Force
 $fileSize = (Get-Item $outputFile).Length / 1MB
 Write-Host "`nSuccess! Created $outputFile ($([math]::Round($fileSize, 2)) MB)" -ForegroundColor Green
 Write-Host "`nTo install:"
-Write-Host "1. Open Zotero 7"
+Write-Host "1. Open Zotero"
 Write-Host "2. Go to Tools -> Add-ons"
 Write-Host "3. Click the gear icon -> 'Install Add-on From File...'"
 Write-Host "4. Select the $outputFile file"